fix build for ALL
now the sharedserver/gameserver/loginserver can run successfully

+++ src/Core/Properties.cpp
@@ -80,7 +80,7 @@ void Properties::load ()
 			continue;
 
 		// key 의 시작문자(white space가 아닌 문자)를 찾는다. 
-		uint key_begin = line.find_first_not_of( WhiteSpaces );
+		size_t key_begin = line.find_first_not_of( WhiteSpaces );
 		
 		// key_begin이 npos 라는 뜻은 그런 문자를 찾지 못했다는 뜻이다.
 		// 즉, 온통 white space 로만 되어 있는 라인이므로 skip 한다.
@@ -90,17 +90,17 @@ void Properties::load ()
 		// key 와 value 를 구분짓는 separator 를 찾는다.
 		// key_end 보다 sep 를 먼저 찾는 이유는 find_last_not_of()를 써서
 		// sep 에서부터 역으로 key_end 를 찾기 위해서이다. ^^;
-		uint sep = line.find( Separator , key_begin );
+		size_t sep = line.find( Separator , key_begin );
 
 		// Separator 를 발견하지 못했을 경우는 파싱 에러로 간주한다.
 		if ( sep == string::npos )
 			throw IOException("missing separator");
 		
 		// sep 에서부터 역으로 key_end 를 찾아나간다.
-		uint key_end = line.find_last_not_of( WhiteSpaces , sep - 1 );
+		size_t key_end = line.find_last_not_of( WhiteSpaces , sep - 1 );
 		
 		// sep 에서부터 value_begin 을 찾는다.
-		uint value_begin = line.find_first_not_of( WhiteSpaces , sep + 1 );
+		size_t value_begin = line.find_first_not_of( WhiteSpaces , sep + 1 );
 		
 		// key 는 있지만 value 가 없는 상태이다. 
 		if ( value_begin == string::npos )
@@ -108,7 +108,7 @@ void Properties::load ()
 		
 		// 맨 끝에서부터 역으로 value_end 를 찾는다. 
 		// ( value_begin 이 있으면 value_end 는 무조건 존재한다.)
-		uint value_end = line.find_last_not_of( WhiteSpaces ); 
+		size_t value_end = line.find_last_not_of( WhiteSpaces ); 
 
 		// key_begin,key_end 와 value_begin,value_end 를 사용해서 
 		// line 의 substring 인 key 와 value 를 생성한다.
