+++ src/Core/SXml.cpp
@@ -438,7 +438,18 @@ void XMLTree::AddAttribute( const string& name, const unsigned int& value, const
 	AddAttribute( name, string(szTemp) );
 }
 
-void XMLTree::AddAttribute( const string& name, const DWORD& value, const bool bHex )
+// void XMLTree::AddAttribute( const string& name, const DWORD& value, const bool bHex )
+// {
+// 	char szTemp[20];
+// 	if( bHex == true )
+// 		itoa( value, szTemp, 16 );
+// 	else
+// 		itoa( value, szTemp, 10 );
+
+// 	AddAttribute( name, string(szTemp) );
+// }
+
+void XMLTree::AddAttribute( const string& name, const unsigned long& value, const bool bHex )
 {
 	char szTemp[20];
 	if( bHex == true )
@@ -527,8 +538,24 @@ XMLTree::GetAttribute( const string& name, unsigned int &value, const bool bHex 
 	return true;
 }
 
+// const bool
+// XMLTree::GetAttribute( const string& name, DWORD &value, const bool bHex )
+// {
+// 	const XMLAttribute *pAttr = GetAttribute( name );
+
+// 	if( pAttr == NULL )
+// 		return false;
+
+// 	if( bHex == true )
+// 		value = pAttr->ToHex();
+// 	else
+// 		value = pAttr->ToInt();
+
+// 	return true;
+// }
+
 const bool
-XMLTree::GetAttribute( const string& name, DWORD &value, const bool bHex )
+XMLTree::GetAttribute( const string& name, unsigned long &value, const bool bHex )
 {
 	const XMLAttribute *pAttr = GetAttribute( name );
