+++ src/Core/VSDateTime.cpp
@@ -188,8 +188,8 @@ VSDate VSDate::fromString(string text)
 	VSDate vsd = VSDate::currentDate();
 
 	// trim white space
-    uint begin = text.find_first_not_of( whiteSpaces );
-    uint end = text.find_last_not_of( whiteSpaces );
+    size_t begin = text.find_first_not_of( whiteSpaces );
+    size_t end = text.find_last_not_of( whiteSpaces );
     text = text.substr( begin , end - begin + 1 );
 
 	/////////////////////////////////////////
@@ -197,8 +197,8 @@ VSDate VSDate::fromString(string text)
 	/////////////////////////////////////////
 	if (text.size() < 8) return vsd;
 
-	uint a=0;
-	uint b=0;
+	size_t a=0;
+	size_t b=0;
 
 	a = text.find_first_of('.', 0  );
 	b = text.find_first_of('.', a+1);
@@ -420,8 +420,8 @@ VSTime VSTime::fromString(string text)
 	VSTime vst = VSTime::currentTime();
 
 	// trim white space
-    uint begin = text.find_first_not_of( whiteSpaces );
-    uint end = text.find_last_not_of( whiteSpaces );
+    size_t begin = text.find_first_not_of( whiteSpaces );
+    size_t end = text.find_last_not_of( whiteSpaces );
     text = text.substr( begin , end - begin + 1 );
 
 	/////////////////////////////////////////
@@ -429,7 +429,7 @@ VSTime VSTime::fromString(string text)
 	/////////////////////////////////////////
 	if (text.size() < 5) return vst;
 
-	uint a=0, b=0, c=0;
+	size_t a=0, b=0, c=0;
 
 	a = text.find_first_of(':', 0  );
 	b = text.find_first_of(':', a+1);
@@ -687,10 +687,10 @@ VSDateTime VSDateTime::currentDateTime()
 VSDateTime VSDateTime::fromString(string text)
 {
 	VSDateTime vsdt = VSDateTime::currentDateTime();
-	uint       a    = 0;
+	size_t       a    = 0;
 
-    uint begin = text.find_first_not_of( whiteSpaces );
-    uint end = text.find_last_not_of( whiteSpaces );
+    size_t begin = text.find_first_not_of( whiteSpaces );
+    size_t end = text.find_last_not_of( whiteSpaces );
     text = text.substr( begin , end - begin + 1 );
 
 	/////////////////////////////////////////
@@ -761,8 +761,8 @@ VSTime VSTime::fromStringHM(string text)
     VSTime vst = VSTime::currentTime();
 
     // trim white space
-    uint begin = text.find_first_not_of( whiteSpaces );
-    uint end = text.find_last_not_of( whiteSpaces );
+    size_t begin = text.find_first_not_of( whiteSpaces );
+    size_t end = text.find_last_not_of( whiteSpaces );
     text = text.substr( begin , end - begin + 1 );
 
     /////////////////////////////////////////
@@ -770,7 +770,7 @@ VSTime VSTime::fromStringHM(string text)
     /////////////////////////////////////////
     if (text.size() < 3) return vst;
 
-    uint a=0;
+    size_t a=0;
 
     a = text.find_first_of(':', 0);
